# docs/api/openApiSpecs.json
{
  "info": {
    "title": "EverLearn: e-learning platform",
    "version": "1.0.1",
    "description": "API documentation"
  },
  "host": "localhost:8000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "swagger": "2.0",
  "paths": {
...
    "/enrollments": {
      "post": {
        "tags": [
          "enrollments"
        ],
        "summary": "To create enrollment.",
        "parameters": [
          {
            "in": "body",
            "name": "enrollment",
            "description": "enrollments data",
            "schema": {
              "$ref": "components/enrollment/req.json#CreateEnrollment"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "successful operation."
          },
          "400": {
            "description": "Bad Request - validation error",
            "schema": {
              "$ref": "components/errorContracts.json#/ValidationErrorResponse"
            }
          },
          "401": {
            "description": "unauthorized access.",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          },
          "403": {
            "description": "forbidden access.",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          }
        }
      }
    },
...
    "/user/register": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "To register new user.",
        "parameters": [
          {
            "in": "body",
            "name": "user",
            "description": "user data",
            "schema": {
              "$ref": "components/user/req.json#RegisterUser"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "successful operation."
          },
          "400": {
            "description": "Bad Request - validation error",
            "schema": {
              "$ref": "components/errorContracts.json#/ValidationErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "To login user.",
        "parameters": [
          {
            "in": "body",
            "name": "user",
            "description": "user data",
            "schema": {
              "$ref": "components/user/req.json#LoginAndTokenGeneration"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation.",
            "schema": {
              "$ref": "components/user/res.json#LoginAndTokenGeneration"
            }
          },
          "400": {
            "description": "Bad Request - validation error",
            "schema": {
              "$ref": "components/errorContracts.json#/ValidationErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          }
        }
      }
    },
    "/user/enrollments": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "To get all courses login user enroll.",
        "responses": {
          "200": {
            "description": "successful operation.",
            "schema": {
              "$ref": "components/user/res.json#LoginAndTokenGeneration"
            }
          },
          "400": {
            "description": "Bad Request - validation error",
            "schema": {
              "$ref": "components/errorContracts.json#/ValidationErrorResponse"
            }
          },
          "401": {
            "description": "unauthorized access.",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          },
          "403": {
            "description": "forbidden access.",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          }
        }
      }
    }
  },
...
  "securityDefinitions": {},
...
}